+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<ma
 ### Added
 * Add derived `total_accel` column to Flysight V2 IMU sensor dataframe, calculated as a vector sum of the `xyz` acceleration components
 * Add `pyflysight.log_utils.locate_log_subdir` helper for resolving child log directory from a given top-level directory
+* Add `pyflysight.log_utils.iter_log_dirs` helper for iterating through child log directories of a given top-level directory
 
 ## [v0.3.0]
 ### Changed